add lifetime deposits

+++ app/src/main/java/com/binancetracker/room/entity/Profit.java
@@ -23,7 +23,7 @@ public class Profit {
     public long tradescount;
 
     @ColumnInfo(name = "deposits")
-    public long deposits;
+    public double deposits;
 
 
 }
